Move everything to jooqflyway
