+++ titus-ext/jooqflyway/build.gradle
@@ -54,123 +54,9 @@ dependencies {
     // Hikari DataSource for connection pooling to the DB
     compile 'com.zaxxer:HikariCP:3.2.0'
 
-    testCompile "org.testcontainers:postgresql:1.15.2"
+    //testCompile "org.testcontainers:postgresql:1.15.2"
     testCompile project(':titus-testkit')
 }
-/*
-def startEmbeddedPostgresDatabase(boolean uselocal) {
-    println 'Starting embedded Postgresql DB...'
-
-    if (uselocal) {
-        jdbcPort = 5432
-        jdbcUrl = 'jdbc:postgresql://localhost:' + jdbcPort + '/postgres'
-        jdbcUser = 'juser'
-        jdbcPassword = 'jpassword'
-    } else {
-        pg = EmbeddedPostgres.start()
-        jdbcUser = 'postgres'
-        jdbcPassword = 'postgres'
-        jdbcPort = pg.getPort()
-        jdbcUrl = 'jdbc:postgresql://localhost:' + pg.getPort() + '/postgres'
-
-        println 'Postgres ' + jdbcUrl
-        dataSource = pg.getPostgresDatabase()
-        // The following code checks we are able to run a test query to the DB
-        Connection connection = dataSource.getConnection()
-        Statement statement = connection.createStatement()
-        ResultSet rs = statement.executeQuery("SELECT 1")
-        assert rs.next()
-        assert rs.getInt(1) == 1
-    }
-    println 'Embedded Postgresql DB started on port: ' + jdbcPort
-}
-
-task startDb {
-    group 'Embedded Postgresql DB'
-    description 'Starts an instance of Embedded Postgresql DB'
-    doFirst {
-        startEmbeddedPostgresDatabase(false)
-        flyway {
-            driver = 'org.postgresql.Driver'
-            url = project.property('jdbcUrl')
-            user = project.property('jdbcUser')
-            password = project.property('jdbcPassword')
-            schemas = ['activity', 'jobactivity']
-            println 'Flyway Configuration for migrating job activity tables'
-            println '=========================================='
-            println 'jdbcUrl:  ' + url
-            println 'user      ' + user
-            println 'password: ' + password
-            println 'schemas:  ' + schemas
-            println '=========================================='
-        }
-        jooq {
-            println 'Setting up JOOQ Configuration'
-            println '============================='
-            version = '3.13.2'
-            cmb(sourceSets.main) {
-                jdbc {
-                    url = project.property('jdbcUrl')
-                    user = project.property('jdbcUser')
-                    password = project.property('jdbcPassword')
-                }
-                generator {
-                    database {
-                        name = 'org.jooq.meta.postgres.PostgresDatabase'
-                        excludes = ''
-                        recordVersionFields = 'rec_version'
-                    }
-                    target { packageName = 'com.netflix.titus.ext.jooqflyway.generated' }
-                    generate {
-                        daos = true
-                        interfaces = true
-                        immutablePojos = false
-                        fluentSetters = true
-                        pojosEqualsAndHashCode = true
-                    }
-                    strategy {
-                        name = null
-                        matchers {
-                            tables {
-                                table {
-                                    pojoClass {
-                                        transform = 'PASCAL'
-                                        // All Jooq Pojo Classess will suffixed with JooqPojo
-                                        expression = '\$0_Jooq_Pojo'
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-task stopDb {
-    group 'Embedded Postgresql DB'
-    description 'Stops a running instance of Embedded Postgresql DB'
-    doLast {
-        // If you need the embedded DB to still be running for debugging code generation or DB migration,
-        // comment the following lines out so the DB is active. Connect to the port number in the output using your local postgres client.
-        // Alternatively, grep the process list (ps) for the word "embedded-pg" to get the postgresql command line arguments.
-        // You will need to manually terminate the postgresql instances running locally.
-        println 'Stopping embedded Postgresql DB running on port ' + jdbcPort
-        if (pg != null) pg.close()
-    }
-}
-
-// This partial configuration is needed so gradle recognizes jooq -> generateCmbJooqSchemaSource task for dev environments
-jooq {
-    println 'setup first'
-    version = '3.13.2'
-    cmb(sourceSets.main) {
-        // ... This will be populated in the startDb method
-    }
-}
-*/
-
 
 // Begin: Jooq code generation using EmbeddedPostgres database and Flyway migration
 ext.dataSource = null
