+++ titus-ext/jooqflyway/gradle/dependency-locks/compileClasspath.lockfile
@@ -16,6 +16,7 @@ com.github.fge:json-schema-core:1.2.5
 com.github.fge:json-schema-validator:2.2.6
 com.github.fge:msg-simple:1.1
 com.github.fge:uri-template:0.9
+com.github.spotbugs:spotbugs-annotations:3.1.12
 com.google.android:annotations:4.1.1.4
 com.google.api.grpc:proto-google-common-protos:1.12.0
 com.google.code.findbugs:annotations:2.0.1
@@ -46,7 +47,7 @@ com.netflix.governator:governator-providers:1.15.11
 com.netflix.governator:governator-servlet:1.15.11
 com.netflix.numerus:numerus:1.1
 com.netflix.spectator:spectator-api:0.59.0
-com.opentable.components:otj-pg-embedded:0.11.4
+com.opentable.components:otj-pg-embedded:0.13.3
 com.squareup.okhttp3:logging-interceptor:3.14.9
 com.squareup.okhttp3:okhttp:3.14.9
 com.squareup.okio:okio:1.17.2
