--- titus-ext/jooqflyway/src/main/java/com/netflix/titus/ext/jooqflyway/jobactivity/JooqProducerDriverComponent.java
@@ -1,4 +1,0 @@
-package com.netflix.titus.ext.jooqflyway.jobactivity;
-
-public class JooqProducerDriverComponent {
-}
